utils/documentai/requirements.txt
google-cloud-bigquery==2.20.0
google-cloud-logging
google-cloud-documentai==2.15.0
google-cloud-storage